+++ sdk/spring/azure-spring-cloud-autoconfigure/src/main/java/com/azure/spring/autoconfigure/jms/AzureServiceBusJMSProperties.java
@@ -4,6 +4,7 @@
 package com.azure.spring.autoconfigure.jms;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.jms.support.QosSettings;
 import org.springframework.util.StringUtils;
 import org.springframework.validation.annotation.Validated;
 
@@ -22,13 +23,15 @@ public class AzureServiceBusJMSProperties {
     private String connectionString;
 
     /**
-     * JMS clientID
+     * JMS clientID. Only works for the bean of topicJmsListenerContainerFactory.
      */
     private String topicClientId;
 
     private int idleTimeout = 1800000;
 
     private String pricingTier;
+
+    private final Listener listener = new Listener();
 
     public String getConnectionString() {
         return connectionString;
@@ -62,6 +65,10 @@ public class AzureServiceBusJMSProperties {
         this.idleTimeout = idleTimeout;
     }
 
+    public Listener getListener() {
+        return listener;
+    }
+
     /**
      * Validate spring.jms.servicebus related properties.
      *
@@ -75,6 +82,79 @@ public class AzureServiceBusJMSProperties {
 
         if (!pricingTier.matches("(?i)premium|standard|basic")) {
             throw new IllegalArgumentException("'spring.jms.servicebus.pricing-tier' is not valid");
+        }
+    }
+
+    /**
+     * Properties to configure {@link org.springframework.jms.annotation.JmsListener} for
+     * {@link org.springframework.jms.config.AbstractJmsListenerContainerFactory}.
+     */
+    public static class Listener {
+
+        /**
+         * Whether the reply destination type is topic. Only works for the bean of topicJmsListenerContainerFactory.
+         */
+        private Boolean replyPubSubDomain;
+
+        /**
+         * Configure the {@link QosSettings} to use when sending a reply.
+         */
+        private QosSettings replyQosSettings;
+
+        /**
+         * Whether to make the subscription durable. Only works for the bean of topicJmsListenerContainerFactory.
+         */
+        private Boolean subscriptionDurable = Boolean.TRUE;
+
+        /**
+         * Whether to make the subscription shared. Only works for the bean of topicJmsListenerContainerFactory.
+         */
+        private Boolean subscriptionShared;
+
+        /**
+         * Specify the phase in which this container should be started and
+         * stopped.
+         */
+        private Integer phase;
+
+        public Boolean isReplyPubSubDomain() {
+            return replyPubSubDomain;
+        }
+
+        public void setReplyPubSubDomain(Boolean replyPubSubDomain) {
+            this.replyPubSubDomain = replyPubSubDomain;
+        }
+
+        public QosSettings getReplyQosSettings() {
+            return replyQosSettings;
+        }
+
+        public void setReplyQosSettings(QosSettings replyQosSettings) {
+            this.replyQosSettings = replyQosSettings;
+        }
+
+        public Boolean isSubscriptionDurable() {
+            return subscriptionDurable;
+        }
+
+        public void setSubscriptionDurable(Boolean subscriptionDurable) {
+            this.subscriptionDurable = subscriptionDurable;
+        }
+
+        public Boolean isSubscriptionShared() {
+            return subscriptionShared;
+        }
+
+        public void setSubscriptionShared(Boolean subscriptionShared) {
+            this.subscriptionShared = subscriptionShared;
+        }
+
+        public Integer getPhase() {
+            return phase;
+        }
+
+        public void setPhase(Integer phase) {
+            this.phase = phase;
         }
     }
 }