+++ sdk/spring/azure-spring-cloud-autoconfigure/src/test/java/com/azure/spring/autoconfigure/jms/NonPremiumServiceBusJMSAutoConfigurationTest.java
@@ -12,18 +12,11 @@ import org.springframework.boot.autoconfigure.AutoConfigurations;
 import org.springframework.boot.autoconfigure.jms.JmsAutoConfiguration;
 import org.springframework.boot.test.context.FilteredClassLoader;
 import org.springframework.boot.test.context.runner.ApplicationContextRunner;
-import org.springframework.jms.core.JmsTemplate;
-
-import javax.jms.ConnectionFactory;
 
 import static com.azure.spring.cloud.autoconfigure.context.AzurePropertyAutoConfiguration.AZURE_PROPERTY_BEAN_NAME;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class NonPremiumServiceBusJMSAutoConfigurationTest {
-
-    private static final String CONNECTION_STRING = "Endpoint=sb://host/;SharedAccessKeyName=sasKeyName;"
-        + "SharedAccessKey=sasKey";
+public class NonPremiumServiceBusJMSAutoConfigurationTest extends AbstractServiceBusJMSAutoConfigurationTest {
 
     @Test
     public void testAzureServiceBusNonPremiumAutoConfiguration() {
@@ -38,12 +31,12 @@ public class NonPremiumServiceBusJMSAutoConfigurationTest {
                      .run(context -> assertThat(context).hasSingleBean(AzureServiceBusJMSProperties.class));
     }
 
+
     @Test
-    public void testAzureServiceBusJMSPropertiesConnectionStringValidation() {
+    public void testWithoutServiceBusJMSNamespace() {
         ApplicationContextRunner contextRunner = getEmptyContextRunner();
-        contextRunner.run(
-            context -> Assertions.assertThrows(IllegalStateException.class,
-                () -> context.getBean(AzureServiceBusJMSProperties.class)));
+        contextRunner.withClassLoader(new FilteredClassLoader(JmsConnectionFactory.class))
+                     .run(context -> assertThat(context).doesNotHaveBean(AzureServiceBusJMSProperties.class));
     }
 
     @Test
@@ -56,44 +49,18 @@ public class NonPremiumServiceBusJMSAutoConfigurationTest {
                          () -> context.getBean(AzureServiceBusJMSProperties.class)));
     }
 
-    @Test
-    public void testWithoutServiceBusJMSNamespace() {
-        ApplicationContextRunner contextRunner = getEmptyContextRunner();
-        contextRunner.withClassLoader(new FilteredClassLoader(JmsConnectionFactory.class))
-                     .run(context -> assertThat(context).doesNotHaveBean(AzureServiceBusJMSProperties.class));
+    @Override
+    protected ApplicationContextRunner getEmptyContextRunner() {
+
+        return new ApplicationContextRunner()
+            .withConfiguration(AutoConfigurations.of(NonPremiumServiceBusJMSAutoConfiguration.class,
+                JmsAutoConfiguration.class, AzurePropertyAutoConfiguration.class))
+            .withPropertyValues(
+                "spring.jms.servicebus.pricing-tier=basic"
+            );
     }
 
-    @Test
-    public void testCachingConnectionFactoryIsAutowired() {
-
-        ApplicationContextRunner contextRunner = getContextRunnerWithProperties();
-
-        contextRunner.run(
-            context -> {
-                assertThat(context).hasSingleBean(ConnectionFactory.class);
-                assertThat(context).hasSingleBean(JmsTemplate.class);
-                ConnectionFactory connectionFactory = context.getBean(ConnectionFactory.class);
-                assertTrue(connectionFactory == context.getBean(JmsTemplate.class).getConnectionFactory());
-            }
-        );
-    }
-
-    @Test
-    public void testAzureServiceBusJMSPropertiesConfigured() {
-
-        ApplicationContextRunner contextRunner = getContextRunnerWithProperties();
-
-        contextRunner.run(
-            context -> {
-                assertThat(context).hasSingleBean(AzureServiceBusJMSProperties.class);
-                assertThat(context.getBean(AzureServiceBusJMSProperties.class).getConnectionString()).isEqualTo(
-                    CONNECTION_STRING);
-                assertThat(context.getBean(AzureServiceBusJMSProperties.class).getTopicClientId()).isEqualTo("cid");
-                assertThat(context.getBean(AzureServiceBusJMSProperties.class).getIdleTimeout()).isEqualTo(123);
-            }
-        );
-    }
-
+/*// TODO (xiada): do we need this test here?
     @Test
     public void testAzurePropertiesConfigured() {
         ApplicationContextRunner contextRunner = getContextRunnerWithProperties();
@@ -107,27 +74,26 @@ public class NonPremiumServiceBusJMSAutoConfigurationTest {
             }
         );
     }
+*/
 
-    private ApplicationContextRunner getEmptyContextRunner() {
-
-        return new ApplicationContextRunner()
-            .withConfiguration(AutoConfigurations.of(NonPremiumServiceBusJMSAutoConfiguration.class,
-                JmsAutoConfiguration.class, AzurePropertyAutoConfiguration.class))
-            .withPropertyValues(
-                "spring.jms.servicebus.pricing-tier=basic"
-            );
-    }
-
-    private ApplicationContextRunner getContextRunnerWithProperties() {
+    @Override
+    protected ApplicationContextRunner getContextRunnerWithProperties() {
 
         return new ApplicationContextRunner()
             .withConfiguration(AutoConfigurations.of(NonPremiumServiceBusJMSAutoConfiguration.class,
                 JmsAutoConfiguration.class, AzurePropertyAutoConfiguration.class))
             .withPropertyValues(
+                "spring.jms.listener.autoStartup=false",
+                "spring.jms.listener.acknowledgeMode=client",
+                "spring.jms.listener.concurrency=2",
+                "spring.jms.listener.receiveTimeout=2s",
+                "spring.jms.listener.maxConcurrency=10",
                 "spring.jms.servicebus.connection-string=" + CONNECTION_STRING,
                 "spring.jms.servicebus.topic-client-id=cid",
                 "spring.jms.servicebus.idle-timeout=123",
                 "spring.jms.servicebus.pricing-tier=basic",
+                "spring.jms.servicebus.listener.reply-pub-sub-domain=false",
+                "spring.jms.servicebus.listener.reply-qos-settings.priority=1",
                 "spring.jms.servicebus.credential.client-secret=for-test-purpose",
                 "spring.jms.servicebus.environment.cloud=AzureGermany",
                 "spring.cloud.azure.credential.client-certificate-password=for-test-purpose",