+++ sdk/spring/scripts/Update-DevOps-Release-WorkItem/Update-DevOps-Release-WorkItems-For-SpringBoot.ps1
@@ -9,7 +9,6 @@ $springtBootReleaseVersion = "3.6.0"
 . ${PSScriptRoot}\Update-DevOps-Release-WorkItem-Util.ps1 $releaseDate $serviceDirectory $springtBootReleaseVersion azure-spring-boot
 . ${PSScriptRoot}\Update-DevOps-Release-WorkItem-Util.ps1 $releaseDate $serviceDirectory $springtBootReleaseVersion azure-spring-boot-starter
 . ${PSScriptRoot}\Update-DevOps-Release-WorkItem-Util.ps1 $releaseDate $serviceDirectory $springtBootReleaseVersion azure-spring-boot-starter-active-directory
-. ${PSScriptRoot}\Update-DevOps-Release-WorkItem-Util.ps1 $releaseDate $serviceDirectory $springtBootReleaseVersion azure-spring-boot-starter-active-directory-b2c
 . ${PSScriptRoot}\Update-DevOps-Release-WorkItem-Util.ps1 $releaseDate $serviceDirectory $springtBootReleaseVersion azure-spring-boot-starter-cosmos
 . ${PSScriptRoot}\Update-DevOps-Release-WorkItem-Util.ps1 $releaseDate $serviceDirectory $springtBootReleaseVersion azure-spring-boot-starter-keyvault-secrets
 . ${PSScriptRoot}\Update-DevOps-Release-WorkItem-Util.ps1 $releaseDate $serviceDirectory $springtBootReleaseVersion azure-spring-boot-starter-servicebus-jms