+++ sdk/core/azure-core-http-okhttp/CHANGELOG.md
@@ -2,6 +2,25 @@
 
 ## 1.8.0-beta.1 (Unreleased)
 
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 1.7.3 (2021-09-07)
+
+### Features Added
+
+- Added support for additional environment configurations. ([#23435](https://github.com/Azure/azure-sdk-for-java/pull/23435))
+
+### Other Changes
+
+#### Dependency Updates
+
+- Upgraded `azure-core` from `1.19.0` to `1.20.0`.
 
 ## 1.7.2 (2021-08-06)