+++ sdk/keyvault/azure-security-keyvault-administration/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 4.1.0-beta.1 (Unreleased)
 
+### Features Added
+- Added support for service version `7.3-preview`.
 
 ## 4.0.1 (2021-07-08)